Use SHA256 hash for branding config
This is part of the FIPS compliance project

Switch brand config to use SHA256 as the primary key rather than MD5.

Implementation notes:
- After the change, existing brand configs will continue to be stored by
md5 and they will continue to operate as before. The brand configs are
designed to be immutable, so updating the existing hashes is probably not
a good idea.
- The migrations which create the predefined themes now use static MD5s
for compatability reasons.
- The database columns continue to be named md5, which isn't ideal.
These could be renamed to 'digest' when this is incorporated upstream
which would mean renaming fields in multiple models.

Test plan
- reset the database and run migrations. the default themes should work
as before
- regression test theme configuration and saving